Extract question outside of route
Question is a reusable component can be reused for our
collaborative rooms.
